flatten selectors
